refactor: split persistence from jetstream streamer
